Relocate TripleUtil methods into Triple and remove TripleUtil class
